+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+<a name="1.1.0"></a>
+### 1.1.0  (2015-07-08)
+
+#### Features
+
+* Uses the W3C local storage integration correctly; rather than using the array
+and dot syntax it will use the setItem, getItem, removeItem and clear methods.
+* Fixed various bugs with the handling of data on windows 7 devices
+
 <a name="1.0.0"></a>
 ### 1.0.0  (2015-06-08)